+++ src/com/github/programmeerpiet1/scavangerhunt/assignments/two/AssignmentTwo.java
@@ -3,14 +3,14 @@ package com.github.programmeerpiet1.scavangerhunt.assignments.two;
 import com.github.programmeerpiet1.scavangerhunt.assignments.Assignment;
 import com.github.programmeerpiet1.scavangerhunt.assignments.domain.Piet;
 
-public class AssignmentTwo extends Assignment<Integer> {
+public class AssignmentTwo extends Assignment<Long> {
 
     public AssignmentTwo(final Piet[] pietjes) {
         super(pietjes);
     }
 
     @Override
-    public Integer run() {
+    public Long run() {
         /*
          * De de opdracht is als volgt:
          * Hoe veel unieke pieten er zijn boven de 30 jaar?
@@ -18,6 +18,6 @@ public class AssignmentTwo extends Assignment<Integer> {
          * note: een piet met dezelfde naam, leeftijd en geslacht is dus NIET uniek.
          * */
 
-        return 0;
+        return 0L;
     }
 }
